# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Storages API",
        "description": "Warehouse management system API",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8080/api"
        }
    ],
    "paths": {
...
        "/storages": {
            "get": {
                "tags": [
                    "Storages"
                ],
                "summary": "List all storages",
                "operationId": "9e00c674911dab8933adeb82c204dc9f",
                "responses": {
                    "200": {
                        "description": "List of storages",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Storage"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Storages"
                ],
                "summary": "Create a new storage",
                "operationId": "6dc3e3855ead1ff134bb62a73f1d00c9",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Yunusabad Warehouse"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Storage created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Storage"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error or name not unique"
                    }
                }
            }
        },
        "/storages/{id}": {
            "put": {
                "tags": [
                    "Storages"
                ],
                "summary": "Update a storage",
                "operationId": "4588a05bb315bc285736bfdc095d4dc2",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Main Warehouse"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Storage updated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Storage"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Storage not found"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Storages"
                ],
                "summary": "Delete a storage",
                "operationId": "9a5077eba3c8482b7e77570dfcd0666c",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Storage deleted"
                    },
                    "404": {
                        "description": "Storage not found"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Storage": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "Main Warehouse"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2026-01-01T00:00:00Z"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2026-01-01T00:00:00Z"
                    }
                },
                "type": "object"
            },
...
        }
    },
...
}